# Apps/StubCustodians/src/SUI.Custodians.API.Client/custodians-openapi-generated.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Custodians API",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "TokenForm",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "client_id": {
                    "type": "string",
                    "additionalProperties": false
                  },
                  "client_secret": {
                    "type": "string",
                    "additionalProperties": false
                  },
                  "scope": {
                    "type": "string",
                    "additionalProperties": false
                  },
                  "grant_type": {
                    "type": "string",
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Fetch/{providerSystemId}/{recordType}/{sui}": {
      "get": {
        "tags": [
          "Fetch"
        ],
        "parameters": [
          {
            "name": "sui",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "recordType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "providerSystemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Fetch/{providerSystemId}/{sui}": {
      "get": {
        "tags": [
          "Fetch"
        ],
        "parameters": [
          {
            "name": "sui",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "providerSystemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEnvelopeOfSuiRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "additionalProperties": false
          },
          "token_type": {
            "type": "string",
            "additionalProperties": false
          },
          "expires_in": {
            "type": "integer",
            "additionalProperties": false,
            "format": "int32"
          },
          "scope": {
            "type": "string",
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
      "ErrorInfo": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "message": {
            "type": "string",
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "FailureInfo": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/FailureKind"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "FailureKind": {
        "type": "integer",
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "title": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "status": {
            "type": "integer",
            "additionalProperties": false,
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "instance": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RecordEnvelopeOfSuiRecord": {
        "required": [
          "schemaUri",
          "payload"
        ],
        "type": "object",
        "properties": {
          "schemaUri": {
            "type": "string",
            "additionalProperties": false,
            "description": "URI of the SuiRecord payload schema",
            "format": "uri",
            "example": "https://schemas.example.gov.uk/sui/SuiRecord"
          },
          "payload": {
            "$ref": "#/components/schemas/SuiRecord"
          }
        },
        "additionalProperties": false
      },
...
      "SuiRecord": {
        "type": "object",
        "additionalProperties": false
      },
...
    }
  },
...
}